+++ packages/blocks/core/js/wordpress/navigation-submenu/test/functionality.blocks-navigation-submenu.e2e.cy.js
@@ -7,6 +7,7 @@ import {
 	appendBlocks,
 	openInserter,
 	setInnerBlock,
+	setParentBlock,
 	redirectToFrontPage,
 } from '@blockera/dev-cypress/js/helpers';
 
@@ -181,7 +182,20 @@ describe(
 			});
 
 			//
-			// 2. Assert front end
+			// 2. Check settings tab
+			//
+			setParentBlock();
+			cy.getByDataTest('settings-tab').click();
+
+			cy.get('.block-editor-block-inspector').within(() => {
+				cy.get('.components-tools-panel-header')
+					.contains('Settings')
+					.scrollIntoView()
+					.should('be.visible');
+			});
+
+			//
+			// 3. Assert front end
 			//
 			savePage();
 			redirectToFrontPage();