+++ packages/blocks/core/CHANGELOG.md
@@ -101,6 +101,7 @@
 - `Login/out` block: support improved.
 - `Media Text` block: support improved.
 - `Navigation Link` block: support improved.
+- `Navigation Submenu` block: support improved.
 
 
 ### Automated Tests
@@ -139,6 +140,7 @@
 - Updated E2E tests to check support of `Login/out` block.
 - Updated E2E tests to check support of `Media Text` block.
 - Updated E2E tests to check support of `Navigation Link` block.
+- Updated E2E tests to check support of `Navigation Submenu` block.
 
 
 ## 1.3.0 (2025-04-27)